requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
et_xmlfile==2.0.0
greenlet==3.2.1
idna==3.10
memory-profiler==0.61.0
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pandas-stubs==2.2.3.250308
psutil==7.0.0
psycopg2==2.9.10
pyarrow==20.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.40
types-pytz==2025.2.0.20250326
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
